Implement tabwidth-agnostic alignment.
Prior to this patch, if the algorithm was applied with a given
indent/tab width and let to produce tabs with that knowledge, the
output would become garbled when viewed on a device with a different
tab width.

This patch removes the requirement to match the device (physical) tab
width to the logical (virtual) tab width assumed by the algorithm.